dz12.1

+++ dz12/dz12.1_crawler_with_state.py
@@ -3,48 +3,46 @@ import os
 from time import time
 from requests_html import HTMLSession
 from reppy.robots import Robots
-from timedeco import time_decorator
+
+domain = ''
+crawled_links = set()
+links_to_crawl = set()
+file_results = None
+
+
+def make_request(url):
+    session = HTMLSession()
+    return session.get(url)
+
+
+def parse_robots(domain):
+    robots_link = f'https://{domain}/robots.txt'
+    return Robots.fetch(robots_link)
+
+
+def init_crawl():
+    global domain, file_results
+    domain = input('Введите домен для краулинга: ')
+    first_link = f'https://{domain}/'
+
+    prepared_response = make_request(first_link)
+    first_link = prepared_response.url
+
+    links_to_crawl.add(first_link)
+
+    file_results = open('checking_results.txt', 'w', encoding='utf-8')
+
 
 if os.path.exists('last_saved_crawl.pickle'):
     continue_last_crawl = input('Do you want to continue last crawl? (y/n): ')
     if continue_last_crawl == 'y':
         with open('last_saved_crawl.pickle', 'rb') as f:
             crawled_links, links_to_crawl, domain = pickle.load(f)
+            file_results = open('checking_results.txt', 'a', encoding='utf-8')
     else:
-        domain = input('Введите домен для краулинга: ')
-        crawled_links = set()
-        links_to_crawl = set()
-        first_link = f'https://{domain}/'
-
-        session = HTMLSession()
-        prepared_response = session.get(first_link)
-        first_link = prepared_response.url
-
-        robots_link = f'https://{domain}/robots.txt'
-        robots = Robots.fetch(robots_link)
-
-        links_to_crawl.add(first_link)
-
+        init_crawl()
 else:
-    domain = input('Введите домен для краулинга: ')
-    crawled_links = set()
-    links_to_crawl = set()
-    first_link = f'https://{domain}/'
-
-    session = HTMLSession()
-    prepared_response = session.get(first_link)
-    first_link = prepared_response.url
-
-    robots_link = f'https://{domain}/robots.txt'
-    robots = Robots.fetch(robots_link)
-
-    links_to_crawl.add(first_link)
-
-session = HTMLSession()
-robots_link = f'https://{domain}/robots.txt'
-robots = Robots.fetch(robots_link)
-
-file_results = open('checking_results.txt', 'w', encoding='utf-8')
+    init_crawl()
 
 while True:
 
@@ -55,7 +53,7 @@ while True:
         url = links_to_crawl.pop()
 
         t1 = time()
-        response = time_decorator(session.get)(url)
+        response = make_request(url)
         t2 = time()
 
         crawled_links.add(url)
@@ -64,7 +62,7 @@ while True:
         for link in response.html.absolute_links:
             if domain not in link:
                 continue
-            if not robots.allowed(link, '*'):
+            if not parse_robots(domain).allowed(link, '*'):
                 continue
             if any(x in link for x in bad_parts):
                 continue
@@ -78,12 +76,13 @@ while True:
         file_results.write(result + '\n')
         file_results.flush()
 
-        os.remove('last_saved_crawl.pickle')
-
     except KeyboardInterrupt:
         with open('last_saved_crawl.pickle', 'wb') as f:
             pickle.dump((crawled_links, links_to_crawl, domain), f)
-        breakpoint()
+        exit()
+
+if os.path.exists('last_saved_crawl.pickle'):
+    os.remove('last_saved_crawl.pickle')
